generate_qrcode_drive/requirements.txt
functions-framework==3.*
qrcode==7.4.2
flask==3.0.2
pillow==10.2.0
google-cloud-storage==2.16.0
google-cloud-iam==2.14.3
google-auth
google-auth-httplib2
google-api-python-client